HW10 Task 3
