+++ functionsHW10.cpp
@@ -1,0 +1,40 @@
+#include "functionsHW10.h"
+
+void swap(int* a, int* b) {
+    int temp = *b;
+    *b = *a;
+    *a = temp;
+
+    std::cout << "Values swapped" << std::endl;
+}
+
+void swap(int& a, int& b) {
+    int temp = b;
+    b = a;
+    a = temp;
+
+    std::cout << "Values swapped" << std::endl;
+}
+
+bool calculateSum(const double* arr, int arrSize, double& sum) {
+    if (arrSize <= 0) {
+        return false;
+    }
+    else {
+        for (int i = 0; i < arrSize; i++) {
+            sum += arr[i];
+        }
+
+        return true;
+    }
+}
+
+bool find(const int* arr, int size, int elem) {
+    for (int i = 0; i < size; i++) {
+        if (*(arr + i) == elem) {
+            return true;
+        }
+    }
+
+    return false;
+}
