+++ functionsHW10.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <iostream>
+
+void swap(int& a, int& b);
+
+void swap(int* a, int* b);
+
+bool calculateSum(const double* arr, int arrSize, double& sum);
+
+bool find(const int* arr, int size, int elem);
